Update README.md
Code Editor App is specially built for web development related coding languages like HTML,CSS,JavaScriot, It has a user-friendly User Interface, A slider to Run the code has been added at the bottmom of the App's screen.It has been added by adding SlideToAct widget in the gradle module,and It also consists of the menu which has several options to make the user code faster by adding structure codes like Add HTML Structure(On clicking which a Html code structure is added in the code section) next option is Add Table(On clicking it adds a sample code for adding tables into our website creatred),Add Form (On clicking ,code to add a simple form with CSS is added as asample which the users can make changes accordingly. Add Image Option,Opens a new submenu with the toast message that "Opening options to add Images". The Subm-menu consists of options like Add Rose Image, Add Car Image, Add Natures Gallery(These will provide a template to the user and they can change it according to thier needs.On clicking any image ,It opens into larger view in the Google Drive.

I am trying to add more features to this so as to make it compatible for all the langauges avalable to code.